# neurosentio-copilot-agent/reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuroSentio Copilot Agent",
    "description": "The standalone backend brain of NeuroSentio Daily Copilot. Manages user profiles, tasks, energy logs, and generates rule-based and LLM-powered daily plans. Designed for neurodivergent-friendly pacing.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/tasks/{task_id}/decompose": {
      "post": {
        "tags": [
          "Decomposition"
        ],
        "summary": "Decompose a task into micro-actions",
        "description": "Breaks a task into tiny, neurodivergent-friendly micro-actions.\n\n- If micro-actions already exist for this task, returns them without regenerating\n  (unless force_regenerate=true).\n- If current_energy < 30, switches to recovery mode (fewer, smaller actions).\n- Uses mock LLM by default \u2014 no API key required.\n- Falls back to rule-based decomposition if LLM fails.",
        "operationId": "decompose_task_endpoint_tasks__task_id__decompose_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDecomposeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDecomposeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/micro-actions": {
      "get": {
        "tags": [
          "Decomposition"
        ],
        "summary": "Get all micro-actions for a task",
        "description": "Returns all micro-actions for the given task, ordered by sort_order.\nScoped to the current user \u2014 another user's tasks return an empty list.",
        "operationId": "get_task_micro_actions_tasks__task_id__micro_actions_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MicroAction"
                  },
                  "title": "Response Get Task Micro Actions Tasks  Task Id  Micro Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/calendar/import/mock": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Import Mock Calendar",
        "description": "Import events with privacy sanitization and post-import back-to-back analysis.",
        "operationId": "import_mock_calendar_calendar_import_mock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List Calendar Events",
        "description": "Get all events overlapping a range, scoped to current user.",
        "operationId": "list_calendar_events_calendar_events_get",
        "parameters": [
          {
            "name": "start_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  },
                  "title": "Response List Calendar Events Calendar Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/day-summary": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Calendar Day Summary",
        "description": "Calculate and return daily summary and scheduling recommendations.",
        "operationId": "get_calendar_day_summary_calendar_day_summary_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date in YYYY-MM-DD format",
              "title": "Date"
            },
            "description": "Date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarDaySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events/{event_id}": {
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Delete Calendar Event",
        "description": "Delete a specific calendar event, scoped to user.",
        "operationId": "delete_calendar_event_calendar_events__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events/{event_id}/title": {
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Redact Calendar Title",
        "description": "Individually purge/redact the title of a calendar event, setting it to '[redacted]'.\nLogs a 'redact_field' privacy action in privacy audit logs.",
        "operationId": "redact_calendar_title_calendar_events__event_id__title_delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CalendarDaySummary": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "high_load_event_count": {
            "type": "integer",
            "title": "High Load Event Count"
          },
          "back_to_back_count": {
            "type": "integer",
            "title": "Back To Back Count"
          },
          "total_busy_minutes": {
            "type": "integer",
            "title": "Total Busy Minutes"
          },
          "total_meeting_minutes": {
            "type": "integer",
            "title": "Total Meeting Minutes"
          },
          "free_blocks": {
            "items": {
              "$ref": "#/components/schemas/FreeBlock"
            },
            "type": "array",
            "title": "Free Blocks"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          }
        },
        "type": "object",
        "required": [
          "date",
          "event_count",
          "high_load_event_count",
          "back_to_back_count",
          "total_busy_minutes",
          "total_meeting_minutes",
          "free_blocks",
          "recommendation"
        ],
        "title": "CalendarDaySummary"
      },
      "CalendarEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "external_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Event Id"
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "attendee_count": {
            "type": "integer",
            "title": "Attendee Count"
          },
          "meeting_type": {
            "type": "string",
            "title": "Meeting Type"
          },
          "load_score": {
            "type": "integer",
            "title": "Load Score"
          },
          "energy_cost": {
            "type": "string",
            "title": "Energy Cost"
          },
          "sensory_cost": {
            "type": "string",
            "title": "Sensory Cost"
          },
          "is_back_to_back": {
            "type": "boolean",
            "title": "Is Back To Back"
          },
          "is_busy": {
            "type": "boolean",
            "title": "Is Busy"
          },
          "raw_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "provider",
          "title",
          "start_time",
          "end_time",
          "attendee_count",
          "meeting_type",
          "load_score",
          "energy_cost",
          "sensory_cost",
          "is_back_to_back",
          "is_busy",
          "created_at",
          "updated_at"
        ],
        "title": "CalendarEvent"
      },
      "CalendarEventCreate": {
        "properties": {
          "external_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Event Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "manual"
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "attendee_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attendee Count",
            "default": 0
          },
          "is_busy": {
            "type": "boolean",
            "title": "Is Busy",
            "default": true
          },
          "raw_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Metadata"
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_time",
          "end_time"
        ],
        "title": "CalendarEventCreate"
      },
      "CalendarImportRequest": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/CalendarEventCreate"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "CalendarImportRequest"
      },
      "CalendarImportResponse": {
        "properties": {
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          },
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CalendarEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "imported_count",
          "updated_count",
          "events"
        ],
        "title": "CalendarImportResponse"
      },
...
      "FreeBlock": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          }
        },
        "type": "object",
        "required": [
          "start_time",
          "end_time",
          "duration_minutes"
        ],
        "title": "FreeBlock"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MicroAction": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "parent_micro_action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Micro Action Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "energy_cost": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Cost"
          },
          "sensory_cost": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory Cost"
          },
          "friction_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Friction Level"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done",
              "snoozed",
              "skipped",
              "deferred"
            ],
            "title": "Status",
            "default": "open"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "MicroAction"
      },
...
      "TaskDecomposeRequest": {
        "properties": {
          "current_energy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Energy",
            "description": "Current battery level 0\u2013100. If omitted, assumes normal mode."
          },
          "sensory_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory State"
          },
          "max_actions": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Max Actions",
            "default": 5
          },
          "force_regenerate": {
            "type": "boolean",
            "title": "Force Regenerate",
            "default": false
          }
        },
        "type": "object",
        "title": "TaskDecomposeRequest"
      },
      "TaskDecomposeResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "micro_actions": {
            "items": {
              "$ref": "#/components/schemas/MicroAction"
            },
            "type": "array",
            "title": "Micro Actions"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "mode",
          "source",
          "message",
          "micro_actions"
        ],
        "title": "TaskDecomposeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}